REFACTOR: Rename file and description of tests
These are operations of basic arithmetics, so let's reflect that in our code
