+++ test/operations.spec.js
@@ -1,8 +1,8 @@
 import { expect } from "chai";
-import { add } from "../src/calculator";
+import { add } from "../src/operations";
 
-describe("Calculator", () => {
-  it("adds two numbers", () => {
+describe("Operations", () => {
+  it("adds two integers", () => {
     expect(add(1, 2)).to.eql(3)
     expect(add(2, 3)).to.eql(5)
   });
